refactor(datasets): add an asynchronous task to generate the datasets' features.
This commit adds an asynchronous task to generate the features of a dataset.
As dataset can be fairly large, storing their features can take up to several dozen of minutes.
Thus it is essential to make this process asynchronous, so it can be executed in the background.

Key changes include:
- Added some utility functions and enums
- Added tasks related fields to the `DatasetVersion` model
- Added a task to generate the dataset feature asynchronously
- Moved the dataset `generate_dataset_category_slug` signal into the clean function instead