clean: remove `map_data` application and partially remove the applications `map_layers` and `shapes`
Those applications are no longer needed, thus they are removed.
`map_data` is completely removed from the project as it did not had any models.
However, `map_layers` and `shapes` are emptied but not yet removed, as the production server needs to be updated beforehand.
Their complete deletion will be performed in a dedicated commit later.